Add svn support
Since voom uses git only to install and update, and GitHub supports svn
for those purposes, it is a simple matter to add support for svn.
The relevant parts of install and update have been broken out into
install(), check_update(), and update() functions.

Remove "-" from grep command; Some versions of grep can't accept the -
parameter, and those that can allow for it to be left off. Should cause
no problems having it gone.

Add search to not install git plugins without git
When the host does not have git, having plugins that rely on git is a
problem. Simple solution is to not install those plugins when git is not
installed.